requirements.txt
numpy~=1.20.1
pypng~=0.0.20